+++ pychron/processing/plotters/options/spectrum.py
@@ -15,15 +15,17 @@
 # ===============================================================================
 
 # ============= enthought library imports =======================
-from traits.api import Str, Int, Property, Bool, Enum, Float, Color, Range, Button
-from traitsui.api import View, Item, Group, HGroup, UItem, EnumEditor, spring, VGroup
+from traits.api import Str, Int, Property, Bool, Enum, Float, Color, Range, Button, HasTraits, Dict, List, on_trait_change
+from traitsui.api import View, Item, Group, HGroup, UItem, EnumEditor, spring, VGroup, InstanceEditor
 
 # ============= standard library imports ========================
 import re
 # ============= local library imports  ==========================
+from pychron.core.helpers.color_generators import colornames
 from pychron.envisage.icon_button_editor import icon_button_editor
 
 from pychron.processing.plotters.options.age import AgeOptions
+from pychron.processing.plotters.options.spectrum_group_options import SpectrumGroupOptions, SpectrumGroupEditor
 from pychron.processing.plotters.options.option import SpectrumPlotOptions
 from pychron.pychron_constants import ERROR_TYPES
 
@@ -41,12 +43,12 @@ class SpectrumOptions(AgeOptions):
 
     include_j_error_in_plateau = Bool(True)
     plateau_age_error_kind = Enum(*ERROR_TYPES)
-    calculate_fixed_plateau = Bool(False)
     # plateau_steps = Property(Str)
     # _plateau_steps = Str
 
-    calculate_fixed_plateau_start = Str
-    calculate_fixed_plateau_end = Str
+    # calculate_fixed_plateau = Bool(False)
+    # calculate_fixed_plateau_start = Str
+    # calculate_fixed_plateau_end = Str
 
     plot_option_name = 'Age'
     display_extract_value = Bool(False)
@@ -64,9 +66,9 @@ class SpectrumOptions(AgeOptions):
     user_envelope_color = Bool
     center_line_style = Enum('solid', 'dash', 'dot dash', 'dot', 'long dash')
     extend_plateau_end_caps = Bool(True)
-    plateau_line_width = Float
-    plateau_line_color = Color
-    user_plateau_line_color = Bool
+    # plateau_line_width = Float
+    # plateau_line_color = Color
+    # user_plateau_line_color = Bool
 
     plateau_method = Enum('Fleck 1977', 'Mahon 1996')
     error_calc_method = Property
@@ -75,6 +77,25 @@ class SpectrumOptions(AgeOptions):
     include_plateau_sample = Bool
     include_plateau_identifier = Bool
 
+    edit_groups_button = Button
+    groups = List
+    group = Property
+
+    def get_group(self, gid):
+        n = len(self.groups)
+        return self.groups[gid % n]
+
+    # handlers
+    @on_trait_change('display_step,display_extract_value')
+    def _handle_labels(self):
+        labels_enabled = self.display_extract_value or self.display_step
+        self.aux_plots[-1].show_labels = labels_enabled
+
+    def _edit_groups_button_fired(self):
+        eg = SpectrumGroupEditor(error_envelopes=self.groups)
+        info = eg.edit_traits()
+        if info.result:
+            self.refresh_plot_needed = True
 
     def _edit_plateau_criteria_fired(self):
         v = View(Item('pc_nsteps', label='Num. Steps', tooltip='Number of contiguous steps'),
@@ -84,6 +105,9 @@ class SpectrumOptions(AgeOptions):
                  title='Edit Plateau Criteria',
                  kind='livemodal')
         self.edit_traits(v)
+
+    def _get_group(self):
+        return self.groups[0]
 
     def _get_error_calc_method(self):
         return self.plateau_age_error_kind
@@ -107,8 +131,8 @@ class SpectrumOptions(AgeOptions):
     # return self._plateau_steps
     #
     # def _set_plateau_steps(self, v):
-    #     if v:
-    #         self._plateau_steps = v
+    # if v:
+    # self._plateau_steps = v
     #
     # def _validate_plateau_steps(self, v):
     #     if plat_regex.match(v):
@@ -122,9 +146,9 @@ class SpectrumOptions(AgeOptions):
     def _get_dump_attrs(self):
         attrs = super(SpectrumOptions, self)._get_dump_attrs()
         return attrs + ['step_nsigma',
-                        'calculate_fixed_plateau',
-                        'calculate_fixed_plateau_start',
-                        'calculate_fixed_plateau_end',
+                        # 'calculate_fixed_plateau',
+                        # 'calculate_fixed_plateau_start',
+                        # 'calculate_fixed_plateau_end',
                         'display_extract_value',
                         'display_step',
                         'display_plateau_info',
@@ -132,11 +156,15 @@ class SpectrumOptions(AgeOptions):
                         'plateau_font_size',
                         'integrated_font_size',
                         'step_label_font_size',
-                        'envelope_alpha',
-                        'user_envelope_color', 'envelope_color',
-                        '_plateau_steps', 'center_line_style',
+                        # 'envelope_alpha',
+                        # 'user_envelope_color', 'envelope_color',
+                        'groups',
+                        # '_plateau_steps',
+                        'center_line_style',
                         'extend_plateau_end_caps',
-                        'plateau_line_width', 'plateau_line_color', 'user_plateau_line_color',
+                        # 'plateau_line_width',
+                        # 'plateau_line_color',
+                        # 'user_plateau_line_color',
                         'include_j_error_in_plateau',
                         'plateau_age_error_kind',
                         'plateau_sig_figs',
@@ -156,37 +184,63 @@ class SpectrumOptions(AgeOptions):
                                       tooltip='Edit Plateau Criteria')),
             Item('center_line_style'),
             Item('extend_plateau_end_caps'),
-            Item('plateau_line_width'),
-            HGroup(UItem('user_plateau_line_color'),
-                   Item('plateau_line_color', enabled_when='user_plateau_line_color')),
+            # Item('plateau_line_width'),
+            # HGroup(UItem('user_plateau_line_color'),
+            #        Item('plateau_line_color', enabled_when='user_plateau_line_color')),
 
             Item('nsigma'),
             Item('plateau_age_error_kind',
                  width=-100,
                  label='Error Type'),
             Item('include_j_error_in_plateau', label='Include J Error'),
-            HGroup(
-                Item('calculate_fixed_plateau',
-                     label='Calc. Plateau',
-                     tooltip='Calculate a plateau over provided steps'),
-                Item('calculate_fixed_plateau_start', label='Start'),
-                Item('calculate_fixed_plateau_end', label='End')),
+            # HGroup(
+            #     Item('calculate_fixed_plateau',
+            #          label='Calc. Plateau',
+            #          tooltip='Calculate a plateau over provided steps'),
+            #     Item('calculate_fixed_plateau_start', label='Start'),
+            #     Item('calculate_fixed_plateau_end', label='End')
+            # ),
             show_border=True,
             label='Plateau')
+
+        # error_grp = VGroup(HGroup(Item('step_nsigma',
+        #                                editor=EnumEditor(values=[1, 2, 3]),
+        #                                tooltip='Set the size of the error envelope in standard deviations',
+        #                                label='N. Sigma'),
+        #                           Item('use_error_envelope_fill', label='Fill')),
+        #                    HGroup(UItem('user_envelope_color'),
+        #                           Item('envelope_color',
+        #                                label='Color',
+        #                                enabled_when='user_envelope_color'),
+        #                           Item('envelope_alpha',
+        #                                label='Opacity',
+        #                                enabled_when='use_error_envelope_fill',
+        #                                tooltip='Set the opacity (alpha-value) for the error envelope')),
+        #                    show_border=True,
+        #                    label='Error Envelope')
+        grp_grp = VGroup(HGroup(UItem('group',
+                                      style='custom',
+                                      editor=InstanceEditor(view='simple_view')),
+                                icon_button_editor('edit_groups_button', 'cog')),
+                         show_border=True,
+                         label='Group Attributes')
 
         error_grp = VGroup(HGroup(Item('step_nsigma',
                                        editor=EnumEditor(values=[1, 2, 3]),
                                        tooltip='Set the size of the error envelope in standard deviations',
-                                       label='N. Sigma'),
-                                  Item('use_error_envelope_fill', label='Fill')),
-                           HGroup(UItem('user_envelope_color'),
-                                  Item('envelope_color',
-                                       label='Color',
-                                       enabled_when='user_envelope_color'),
-                                  Item('envelope_alpha',
-                                       label='Opacity',
-                                       enabled_when='use_error_envelope_fill',
-                                       tooltip='Set the opacity (alpha-value) for the error envelope')),
+                                       label='N. Sigma')),
+                           # HGroup(UItem('error_envelope',
+                           #              style='custom',
+                           #              editor=InstanceEditor(view='simple_view')),
+                           #        icon_button_editor('edit_envelopes_button', 'cog')),
+                           # HGroup(UItem('user_envelope_color'),
+                           #        Item('envelope_color',
+                           #             label='Color',
+                           #             enabled_when='user_envelope_color'),
+                           #        Item('envelope_alpha',
+                           #             label='Opacity',
+                           #             enabled_when='use_error_envelope_fill',
+                           #             tooltip='Set the opacity (alpha-value) for the error envelope')),
                            show_border=True,
                            label='Error Envelope')
 
@@ -229,6 +283,7 @@ class SpectrumOptions(AgeOptions):
                             label='Display')
         g = Group(
             self._get_title_group(),
+            grp_grp,
             plat_grp,
             error_grp,
             display_grp,
@@ -251,13 +306,20 @@ class SpectrumOptions(AgeOptions):
                                            'plateau_line_color',
                                            'plateau_font_size',
                                            'plateau_sig_figs',
-                                           'calculate_fixed_plateau',
-                                           'calculate_fixed_plateau_start',
-                                           'calculate_fixed_plateau_end',
+                                           # 'calculate_fixed_plateau',
+                                           # 'calculate_fixed_plateau_start',
+                                           # 'calculate_fixed_plateau_end',
                                            'pc_nsteps',
                                            'pc_gas_fraction'))
 
         self._set_defaults(yd, 'integrated', ('integrated_font_size',
                                               'integrated_sig_figs',))
+        self._set_defaults(yd, 'labels', ('display_step',
+                                          'display_extract_value',
+                                          'step_label_font_size'))
 
+    def _groups_default(self):
+        return [SpectrumGroupOptions(color=ci,
+                                     line_color=ci,
+                                     group_id=i) for i, ci in enumerate(colornames)]
 # ============= EOF =============================================