+++ pychron/processing/plotters/spectrum/tools.py
@@ -59,21 +59,20 @@ class SpectrumTool(InfoInspector, BasePlateauOverlay):
         if ndx < len(ys):
             yd = ys[ndx]
 
-
-            e = comp.errors[ndx*2] * self.nsigma
+            e = comp.errors[ndx * 2] * self.nsigma
             yl, yu = comp.y_mapper.map_screen(array([yd - e, yd + e]))
             # print ys
             # print ndx, yd, yd+e, yd-e, yu, yl, screen_pt[1], self.nsigma
-            if yu-yl<1:
-                yu+=1
-                yl-=1
+            if yu - yl < 1:
+                yu += 1
+                yl -= 1
 
             if yl < screen_pt[1] < yu:
                 return ndx
 
     # def normal_mouse_move(self, event):
     # xy=event.x, event.y
-    #     pos=self.hittest(xy)
+    # pos=self.hittest(xy)
     #     if pos is not None:
     #     # if isinstance(pos, tuple):
     #         self.current_position = pos
@@ -100,7 +99,7 @@ class SpectrumTool(InfoInspector, BasePlateauOverlay):
         idx = self.current_position
         comp = self.component
 
-        e = comp.errors[idx*2]
+        e = comp.errors[idx * 2]
         ys = comp.value.get_data()[::2]
         v = ys[idx]
 
@@ -110,8 +109,8 @@ class SpectrumTool(InfoInspector, BasePlateauOverlay):
         return ['RunID={}'.format(an.record_id),
                 'Tag={}'.format(an.tag),
                 'Status={}'.format(an.status_text),
-                u'{}={} {}{} (1\u03c3)'.format(comp.container.y_axis.title, floatfmt(v),PLUSMINUS,
-                                           floatfmt(e)),
+                u'{}={} {}{} (1\u03c3)'.format(comp.container.y_axis.title, floatfmt(v), PLUSMINUS,
+                                               floatfmt(e)),
                 'Cumulative. Ar39={}-{}'.format(floatfmt(low_c),
                                                 floatfmt(self.cumulative39s[idx]))]
 
@@ -141,7 +140,7 @@ class SpectrumInspectorOverlay(InfoOverlay):
     # tool =Any
     # @on_trait_change('tool:current_section')
     # def handle(self, new):
-    #     if new>=0:
+    # if new>=0:
     #         self.visible=True
     #     else:
     #         self.visible=False
@@ -175,11 +174,19 @@ class SpectrumErrorOverlay(AbstractOverlay):
             es = es.reshape(n / 2, 2)
 
             if self.use_fill:
-                alpha = self.alpha
+                alpha = self.alpha * 0.01
                 func = gc.fill_path
             else:
                 alpha = 1.0
                 func = gc.stroke_path
+
+            color = map(lambda x: x / 255., self.user_color.toTuple())
+
+            color = color[0], color[1], color[2], alpha
+            if abs(alpha-0.3) < 0.1:
+                selection_color = (0.75, 0, 0)
+            else:
+                selection_color = color[0], color[1], color[2], 0.3
 
             for i, ((xa, xb), (ya, yb), (ea, eb)) in enumerate(zip(xs, ys, es)):
                 ea *= self.nsigma
@@ -193,22 +200,25 @@ class SpectrumErrorOverlay(AbstractOverlay):
                 y = p1[1]
                 w = p3[0] - p1[0]
                 h = p2[1] - p1[1]
-                if i in sels:
-                    gc.set_fill_color((0.75, 0, 0))
-                    gc.set_stroke_color((0.75, 0, 0))
-                else:
-                    if self.use_user_color:
-                        c = map(lambda x: x/255., self.user_color.toTuple())
 
-                    else:
-                        c = comp.color
-                        if isinstance(c, str):
-                            c = color_table[c]
+                c = selection_color if i in sels else color
+                # if i in sels:
 
-                    c = c[0], c[1], c[2], alpha
+                # gc.set_fill_color((0.75, 0, 0))
+                # gc.set_stroke_color((0.75, 0, 0))
+                # else:
+                # if self.use_user_color:
+                # c = map(lambda x: x/255., self.user_color.toTuple())
 
-                    gc.set_fill_color(c)
-                    gc.set_stroke_color(c)
+                # else:
+                # c = comp.color
+                #     if isinstance(c, str):
+                #         c = color_table[c]
+
+                # c = c[0], c[1], c[2], alpha
+
+                gc.set_fill_color(c)
+                gc.set_stroke_color(c)
 
                 gc.rect(x, y, w, h)
                 func()
@@ -216,9 +226,14 @@ class SpectrumErrorOverlay(AbstractOverlay):
 
 
 class PlateauTool(DragTool):
+    def normal_mouse_move(self, event):
+        if self.is_draggable(event.x, event.y):
+            event.window.set_pointer('hand')
+        else:
+            event.window.set_pointer('arrow')
     # def normal_mouse_move(self, event):
     # if self.is_draggable(event.x, event.y):
-    #         event.handled = True
+    # event.handled = True
     #
     # def normal_left_down(self, event):
     #     if self.is_draggable(event.x, event.y):
@@ -334,7 +349,7 @@ class PlateauOverlay(BasePlateauOverlay):
 
             # if y1 < y and y2<y:
             # gc.lines([(x1, y1+5), (x1, y + 10)])
-            #     gc.lines([(x2, y2+5), (x2, y + 10)])
+            # gc.lines([(x2, y2+5), (x2, y + 10)])
             # elif y1> y and y2>y:
             #     gc.lines([(x1, y - 10),(x1, y1 + 5)])
             #     gc.lines([(x2, y - 10),(x2, y2 + 5)])
@@ -398,7 +413,7 @@ class PlateauOverlay(BasePlateauOverlay):
 
             p = PlotLabel(text=self.info_txt,
                           font='modern {}'.format(self.label_font_size),
-                          color = self.line_color,
+                          color=self.line_color,
                           hjustify=hjustify,
                           x=x,
                           y=oy + 15)