+++ CHANGELOG.md
@@ -22,6 +22,8 @@ CHANGELOG
 * added ability to control plateau envelope/line color
 * spectrum labels match envelope color
 * display number of analyses as n/m if n != m. n = # of valid analyses, m = total # of analyses
+* added group attributes to spectrum options. allows you to set color for group and calculate a fixed plateau for a specific group
+
 
 ### Bug Fixes ###
 * fixed plotting ideograms of attributes other than age